requirements.txt
python-dotenv
tiktoken
pandas
tqdm
chromadb
sentence-transformers
requests
torch